Fix server Footer importing music destinations from a client module.
